+++ src/ColorBar.ts
