+++ src/RGBConverter.ts
